requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.8.0
eventlet==0.33.3
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
waitress==2.1.2
